+++ CHANGELOG.md
@@ -11,13 +11,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - TBD
 ### Added
 - Added new inline key `F90_RCONST_USE` in `src/gdata.h` and `src/scanner.c`
+- Added documentation about `F90_RCONST_USE` for ReadTheDocs
 
 ### Changed
 - Updated `Update_RCONST` to use `Y` instead of `C` to account for updated variable species concentrations
 - Updated C-I tests to print the compiler versions that are used
 - Updated routine `GenerateUpdateRconst` to manually write the `SUBROUTINE` and `END SUBROUTINE` lines (F90 only)
 - Updated routine `GenerateUpdateRconst` to inline code from `#INLINE F90_RCONST_USE` before any other F90 variable declarations or statements
-- Updated `.gitignore` to ignore executables in the MCM example folders
+- Updated `.gitignore` to ignore all executable files
+- Changed `Begin INLINED RCONST - F90 USE STATEMENTS` to `Begin inlined code from F90_RCONST_USE` in `src/gen.c`
+- Changed `Begin INLINED RCONST` to `Begin inlined code from F90_RCONST` in `src/gen.c`
 
 ### Fixed
 - Added `char* rootFileName` to functions and function prototypes for `Use_C`, `Use_F`, `Use_F90`, `Use_MATLAB`, and `Generate`